test: move image registry unit tests to its own package
