+++ pkg/handlers/aws/mutation/metapatch_handler_test.go
@@ -10,8 +10,6 @@ import (
 
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/common/pkg/capi/clustertopology/handlers/mutation"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/clusterconfig"
-	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/mutation/imageregistries"
-	imageregistrycredentialstests "github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/mutation/imageregistries/credentials/tests"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/mutation/kubernetesimagerepository"
 	kubernetesimagerepositorytests "github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/mutation/kubernetesimagerepository/tests"
 	"github.com/d2iq-labs/cluster-api-runtime-extensions-nutanix/pkg/handlers/generic/mutation/mirrors"
@@ -36,14 +34,6 @@ func TestGeneratePatches(t *testing.T) {
 		metaPatchGeneratorFunc(mgr),
 		clusterconfig.MetaVariableName,
 		kubernetesimagerepository.VariableName,
-	)
-
-	imageregistrycredentialstests.TestGeneratePatches(
-		t,
-		metaPatchGeneratorFunc(mgr),
-		mgr.GetClient(),
-		clusterconfig.MetaVariableName,
-		imageregistries.VariableName,
 	)
 
 	globalimageregistrymirrortests.TestGeneratePatches(
